DCmC 0.7.4 with cool new stuff, issue #122 closed
