+++ include/tower/tower_atk.h
@@ -1,35 +1,12 @@
 //TITLE: TOWER_ATK_H
 //PROJECT: DON´T CRUSH MY CASTLE
 //AUTHOR: Andrés Ortiz
-//VERSION: 0.7
+//VERSION: 0.7.4
 //DESCRIPTION: defines the attack of buildings
-#ifndef TOWER_ATK_H
-#define TOWER_ATK_H
+#ifndef TOWER_ATK
+#define TOWER_ATK
 
-#include "al_anim.h"
-#include "game_object.h"
-enum atk_type {shoot_atk,explosion_atk,continuous_atk};
-
-//defines the atributes of an attack
-struct atk_attributes {
-    ALLEGRO_BITMAP *bitmap; //attack bitmap
-    al_anim collision_anim; //collision animation (explosion) or continuous
-    unsigned int damage; //damage of the attack
-    unsigned int range; //range of damage in pixels
-    double delay; //delay (in seconds) between attacks (except continuous)
-    atk_type type;
-    double speed; //speed of attack (in pixels/second)
-    //string name ¿?
-    //METHODS
-    atk_attributes();
-    atk_attributes(ALLEGRO_BITMAP *bitmap,al_anim collision_anim,unsigned int damage,unsigned int range,unsigned int delay,double speed,atk_type type=shoot_atk);
-    ~atk_attributes();
-    //clear data (dont destroy animations)
-    void clear();
-    //destroy all bitmaps and clear data
-    void destroy();
-    bool check() const;
-};
+#include "tower_atk_attributes.h"
 
 //defines an instance of attack
 class tower_atk : public game_object {
@@ -63,9 +40,9 @@ public:
     //return true after collision
     bool hit() const;
     //updates the attack,updating its position and colliding if necessary
-    void update();
+    void virtual update();
     //draws the attack
-    void draw() const;
+    void virtual draw() const;
     //check class, return false if problem
     bool check() const;
 private:
