+++ makefile
@@ -29,13 +29,13 @@ MAINDIR=$(SDIR)/main
 _INC=$(UTILSDIR) $(MAPDIR) $(ENEMYDIR) $(TOWERDIR) $(CONTROLLERDIR)
 I_INC=$(patsubst %,-I %,$(_INC))
 
-_AL_UTILS=al_anim.cpp al_utils.cpp debug_log.cpp input_handler.cpp game_object.cpp text_handler.cpp tinyxml2.cpp
+_AL_UTILS=al_anim.cpp al_utils.cpp debug_log.cpp input_handler.cpp game_object_attributes.cpp game_object.cpp text_handler.cpp tinyxml2.cpp
 AL_UTILS_O=$(patsubst %,$(ODIR)/%,$(_AL_UTILS:.cpp=.o))
 _MAP=tile.cpp tileset.cpp tilemap.cpp
 MAP_O=$(patsubst %,$(ODIR)/%,$(_MAP:.cpp=.o))
-_ENEMY=enemy.cpp enemy_set.cpp
+_ENEMY=enemy_attributes.cpp enemy.cpp enemy_set.cpp
 ENEMY_O=$(patsubst %,$(ODIR)/%,$(_ENEMY:.cpp=.o))
-_TOWER=tower_atk.cpp tower.cpp tower_set.cpp
+_TOWER=tower_atk_attributes.cpp tower_atk.cpp tower_attributes.cpp tower.cpp tower_set.cpp
 TOWER_O=$(patsubst %,$(ODIR)/%,$(_TOWER:.cpp=.o))
 _CONTROLLER=game_objects.cpp player_controller.cpp game_master.cpp player.cpp
 CONTROLLER_O=$(patsubst %,$(ODIR)/%,$(_CONTROLLER:.cpp=.o))
