+++ src/utilities/game_object_attributes.cpp
@@ -1,0 +1,11 @@
+//TITLE: GAME_OBJECT_ATTRIBUTES_CPP
+//PROJECT: DON´T CRUSH MY CASTLE
+//AUTHOR: Andrés Ortiz
+//VERSION: 0.7.4
+//DESCRIPTION: defines a generic object attributes as a base class
+
+#include "game_object_attributes.h"
+
+bool game_object_attributes::check() const {
+    return false;
+}
